# do type checks when register, not when resolving